+++ src/models/fields/base-field-configuration.ts
@@ -11,12 +11,21 @@ const commonBaseFieldConfigurationSchema = Joi.object().keys({
   showIf: Joi.object().keys({
     key: Joi.string().required(),
     value: Joi.alternatives(
-      Joi.string().strict(),
+      Joi.string().allow('').strict(),
       Joi.number().strict(),
       Joi.boolean().strict(),
     ).required(),
   }).description('Show this field if other field with provided key is set to provided value. Referenced field must exists already'),
   hint: stringOrTranslationsSchema.allow('').description('As shown near the input field'),
+  // the below are needed to validate objects that do not need these props, but
+  // have them set at undefined
+  prefix: Joi.any().forbidden(),
+  suffix: Joi.any().forbidden(),
+  defaultValue: Joi.any().forbidden(),
+  type: Joi.any().forbidden(),
+  regex: Joi.any().forbidden(),
+  lowerbound: Joi.any().forbidden(),
+  upperbound: Joi.any().forbidden(),
 });
 
 const prefixesMixin = {
@@ -41,7 +50,7 @@ const getBaseFieldConfigurationSchema = (
     defaultValue: stringOrTranslationsSchema.allow('').default(''),
     valueOptions: Joi.array().min(1).items(Joi.object().keys({
       text: stringOrTranslationsSchema.required(),
-      value: Joi.string().required(),
+      value: Joi.string().allow('').required(),
     })).required(),
     inputType: Joi.string().valid('select').default('select'),
     ...prefixesMixin,
@@ -51,7 +60,7 @@ const getBaseFieldConfigurationSchema = (
     defaultValue: stringOrTranslationsSchema.allow('').default(''),
     valueOptions: Joi.array().min(1).items(Joi.object().keys({
       text: stringOrTranslationsSchema.required(),
-      value: Joi.string().required(),
+      value: Joi.string().allow('').required(),
     })).required(),
     inputType: Joi.string().valid('radio').required(),
   }),
@@ -66,7 +75,7 @@ const getBaseFieldConfigurationSchema = (
     ...prefixesMixin,
   }),
   commonSchema.keys({
-    type: Joi.string().valid('number').required(),
+    type: Joi.string().valid('number').default('number'),
     defaultValue: Joi.number().default(0),
     valueOptions: Joi.array().min(1).items(Joi.object().keys({
       text: stringOrTranslationsSchema.required(),
@@ -76,7 +85,7 @@ const getBaseFieldConfigurationSchema = (
     ...prefixesMixin,
   }),
   commonSchema.keys({
-    type: Joi.string().valid('number').required(),
+    type: Joi.string().valid('number').default('number'),
     defaultValue: Joi.number().default(0),
     valueOptions: Joi.array().min(1).items(Joi.object().keys({
       text: stringOrTranslationsSchema.required(),
@@ -121,7 +130,14 @@ const getBaseFieldConfigurationSchema = (
     inputType: Joi.string().valid('switch', 'checkbox').default('checkbox'),
   }),
   commonSchema.keys({
-    type: Joi.string().valid('boolean').required(),
+    type: Joi.string().valid('boolean').default('boolean'),
+    defaultValue: Joi.boolean().default(false),
+    valueOptions: Joi.array().items(Joi.any()).length(0).allow(null)
+      .default(null),
+    inputType: Joi.string().valid('switch', 'checkbox').required(),
+  }),
+  commonSchema.keys({
+    type: Joi.string().valid('boolean').default('boolean'),
     defaultValue: Joi.boolean().default(false),
     valueOptions: Joi.array().min(1).items(Joi.object().keys({
       text: stringOrTranslationsSchema.required(),
@@ -131,7 +147,7 @@ const getBaseFieldConfigurationSchema = (
     ...prefixesMixin,
   }),
   commonSchema.keys({
-    type: Joi.string().valid('boolean').required(),
+    type: Joi.string().valid('boolean').default('boolean'),
     defaultValue: Joi.boolean().default(false),
     valueOptions: Joi.array().min(1).items(Joi.object().keys({
       text: stringOrTranslationsSchema.required(),